Update Google OpenID Connect endpoints
Tuleap comes with a preconfiguration for Google but the
endpoints currently used are deprecated and are not compliant with
the OpenID Connect Core standard.
This contribution updates the Google endpoints so that the plugin
still works with Google as an OIDC provider.

request #12235: Drop usage of InoOicClient library in the OpenIDConnect client plugin

Change-Id: Ia1625b0e84f37851e823c5f4fec861454fc89eff