Migrate TextField to VueJS
Part of story #13453: finish drag'n drop images in tracker text areas

Another patch is needed for TestManagement.

How to test:
- rebuild kanban and scrum planning.
- No functional changes expected. Text fields in the Artifact modal
behave correctly with regards to:
  - HTML / Text mode
  - Required
  - Disabled (read-only permission)

---
Why?
Because the AngularJS integration lib we were using, ng.ckeditor, does
not let us access the editor instance, which is needed to handle drag'n
drop of images.
Since we are going to eventually migrate the modal to Vue anyway, we did
not try to fork, fix or rewrite ng.ckeditor. Instead, we wrote a
lightweight integration for Vue.

Change-Id: I6a93447e1c0e5a8aec989fecd519dcf78a034657