request #11133 Can't use "Start Scrum" if TTM is already activated
"Start Scrum" is displaying that message
"The short name 'bugs' is already used. Please use another one."
if TTM is already activated on a project, while it is not the case if
Scrum has been set up before TTM.

This commit introduces extra options for plugin to have custom behviour
while importing. The new Tracker extra option allow to reuse one or more tracker
if they already exist on the target platform.

Change-Id: I88984145ed8dabbf864a6cf0320fdf8dc1693ff8