OIDC authentication URLs are now generated without the help of an external lib
Tuleap now manages directly the creation of the authentication URLs.

This contribution reduces the dependency to an external library, the goal is
to remove it completly to be able to drop completly the non compatible code with
PHP 7.2 code that is currently used.

request #12235: Drop usage of InoOicClient library in the OpenIDConnect client plugin

Change-Id: I210a6dacdb9c0981eab0ab01bd5ef81ecdf37056