Authentication response are processed without the help of an external lib
Tuleap now manages directly the authentication response from the
authentication endpoint.
This contribution reduces the dependency to an external library,
the goal is to remove it completly to be able to drop completly
the non compatible with PHP 7.2 code that is currently used.

request #12235: Drop usage of InoOicClient library in the OpenIDConnect client plugin

Change-Id: I77441b2e50cf074c80a3a60ce939a788f127df42