The OpenID Connect plugin works without an external OIDC library
This contribution removes the last bit of the InoOicClient library
to make Tuleap completly independent of it.
Consequently, the remaining warnings on PHP 7 that was caused by
incoherent method redefinitions in the Tuleap code are no more
present and the plugin is fully compatible with PHP 7.

request #12235: Drop usage of InoOicClient library in the OpenIDConnect client plugin

Change-Id: I390a8835ffffedfe5e9c7aae5e0c86e3536d3ca7