Cache locally the last changeset of an artifact
Currently when the last changeste of artifact is queried from the
Tracker_Artifact, a cache might be used if the object has loaded
all the changesets of the changeset. However if the only thing you
only ask for is the last changeset, it will be retrieved from the
DB each time. This typically happens when you are rendering a tracker
report table.

This is part of request #12245: Displaying an artifact with a lot of artifact
links is slow

Change-Id: I63ef30275e540e90fb211058ae5253e587761408