Use the namespace Tuleap\User everywhere
Depending on the area the namespace Tuleap\User or Tuleap\user where
used. It is a bit painful to work with because autoloader generated by
Composer are case sensitive.
The folder src/common/user is renamed to ensure the PSR-4 compliance so
developers does not have to dump the autoload each time a new class is
added.

This is part of story #12266: use an API key rather than a login / password

Change-Id: I496a5df8dc32898f456d23a9b866ac28c9ba29bf