Retrieving a set of commits via the REST API is fast again
The addition of the status and of the author as a Tuleap user
has slow down tremendously the retrieval of commit representations.
Consequently, the REST routes allowing to retrieve tags, branches or
the commits of a pull request can be too slow to be usable in the web UI.

This contribution solves the issue on working on the two root causes:
 * an index has been added on the user table to be able to quickly
find users matching an email.
 * the number of SQL queries has been reduced. For getting the representations
of a set of n commits, 2*n queries were necessary now only 2 queries are needed.

This is part of:
 - story #10414: have a branch/tag selector
 - story #11664: display list of commits of a pull request

Change-Id: I36e1b3ed36545772f43edc8daf6ebeccc6fc438b