Convert blob view to burning parrot
Part of story #10416: burning parrot gloss on gitphp views

Display the content of a file in a repository.

If it is an image, then it is displayed inline.
If it is a known code (ex: .php, .js), then it is syntax highlighted.
If it is not known code (ex: text file), then it is simply displayed.
If it is binary file, then it is displayed as garbage, just like before. (If it
needs to be fixed, it should be done in a dedicated changeset)

We can acess to the blame, the history or the raw (plain) version of the file.

No change if $git_repository_bp is not set.

Change-Id: I742e9a6fd8c47ddbaad48cc48f9d52a4295c910f