Add minimal support of commit signature
Commit signature does not pollute the commit message anymore.
Tuleap backend is capable to extract the signature and seperate it
from the rest of the commit message.
For now as nothing smart can be done with it, the signature is not displayed
in the web UI. However, commit representation in the REST routes has the
information if available.

Management of the signature of annoted tags could come in a dedicated
contribution if needed.

This is part of story #10416: burning parrot gloss on gitphp views

Change-Id: I89171c4f2172736cbd138a0486b18039fe6ffac8