Use sodium_(bin2hex|hex2bin) instead of paragonie/constant_time_encoding
Constant time encoding/decoding to/from hexadecimal can be done with
libsodium, let's leverage that. It's one less lib and with libsodium in
PHP 7.2 it will be one less side channel potential issue caused by a pure
PHP implementation.

This is part of story #12266: use an API key rather than a login / password

Change-Id: I9903fcf3e2db9b3471c1c67a299f4a256b3aa4cf