Update last access date of an access key according to the last_access_resolution setting
The goal is to limit a write hotspot in the DB when a user queries intensively
the REST API.
The last access resolution has been lowered to 1 hour which seem a good
compromise.
The last_access_resolution setting is only taken into account if the IP
of the user have not changed, otherwise the last access values are updated.

This is part of story #12266: use an API key rather than a login / password

Change-Id: I7762589ae7b18cb037a60b569b20be4652b98cab