Branch/tag selector should keep context
Part of story #10416: burning parrot gloss on gitphp views

Go to a folder or a file. With the selector choose another branch/tag. You
should be redirected to the same folder or file, but on the selected
branch/tag. If the resource does not exist on the new ref, then an error
message is shown.

The same for blame/history. If file has only one commit for the selected ref,
then blame & history don't show empty page.

Change-Id: Id2f10ddfea25eed9eb3fb6c5cc151f7700ae0643