request #12837: Jenkins build history widget generates warnings in PHP 7
In order to remove the warnings, this contribution replace SimplePie
with Zend\Feed.

The version of SimplePie currently bundled with Tuleap is outdated.
Since we already have a dependency to the Zend framework module it makes
sense to reuse that and use Zend\Feed.

Zend\Feed\Writer could also be used to replace our existing self made
and likely broken implementation of RSS feeds.

Change-Id: Ib02d3a5d1c0697401cc7a977294db367b5e36089