Proper whitespace in commit title
Part of story #10416: burning parrot gloss on gitphp views

Go to a commit that contains a cross ref (eg. "This closes art #1: blah")
There should not be whitespace issue around the link <art #1>.

Furthermore if your commit message contains a very long line, it should be
automatically wrapped and should not go outside the pane.

Bonus: remove not needed empty line at the beginning of the description.

Change-Id: Ic3e6ff9fc1e3e9b0b0713294a5fad83d85f5949b