Project owner can be updated via the REST API
The user to set as the project owner can referenced the same
way than the REST route allowing to update the user groups.

Data model of the project certification was wrong, multiple
project owners was allowed per project but the whole feature is planned
for only project owner per project.
The model has been updated but a forgeupgrade can not be provided as the
operation can not be done without data loss. The table must either be
completely dropped and created again or manually updated. The plugin is
not available in the repositories yet so this only impact a few integrators.

This is part of story #12161 have a project owner at project creation

Change-Id: I6e0498bc6de17ffa5f366f3abc4ee33766f01384