Use PrismJS to highlight syntax
Part of request #12394: Blob view is slow

In blob and blame view, use PrismJS instead of GeSHi for syntax highlighting.

This avoid parsing code on the server and sending a big DOM to the client.

On Tuleap 10.6, Tracker.class.php blob display was taking approximatively 10s;
now it takes 2s wich is much better.

Change-Id: Iddca8d60e9f53efc0533095eb3129276c439e5cb