request #12387: Extraction of references followed by an HTML entity is broken
The reference extractor should not try to partially capture a HTML entity
at the end of the reference as it breaks the display of the value represented
by the HTML entity.

Decimal and hexadecimal HTML entities are fully managed. The only named
HTML entity currently managed is &quot;. The match of named HTML entity
should only catch exiting references [0] to avoid too much false positive.
As it was not the source of the present issue only &quot; has been
added, others can be added later if needed. Trying to catch all existing
named references is probably not worth the trouble of maintaining an up
to date list.

[0] https://html.spec.whatwg.org/multipage/named-characters.html#named-character-references

Change-Id: Ie974e47d659a4d546151aa12972c48976666297e