request #12399: Broken navigation on git repository
Plain button should be renamed as Download so that it better reflects the action.
Download button should not be glued to the button bar.
Download should be present in commit view as well, to download the commit as a patch.
On commit view, we should be able to browse source at this specific commit.
On blame view, the file path should be displayed.

Change-Id: I562bcffe67d0c7cca750efb173a0f2d39b84f17f