Remove unneeded call to setlocale(LC_TIME)
This fix request #12433: Empty date when there are accents

When language is loaded, setlocale is already called by BaseLanguageFactory,
therefore it is useless to call it again. Furthermore it can lead to issue on
some environments where locale is not properly set with utf8 encoding.

On your dev environment, you may need to issue the following command so that
you have same locales than prod servers:

 $ yum reinstall glibc-common

Then, switch your preference to french.
Go to a user profile that was registered in august (in french: août)
You should see the date instead of a blank string.
August timestamp example if needed: 1503288092

The same in system events admin interface.

Change-Id: I69801a5ddb0fad071fa6f977900e0bf83193632a