Reorganize SVN Redis cache to be able to revoke cache per project
User seen per project are now stored in a zset so that the cache of
a project can be invalidated. The implementation is based upon the
recommendation/trick mentionned here [0] since it is not possible
to set an expiration on a member of a set.

No functionnal changes are expected from this contribution, the
invalidation of the cache when a project has a status change will
done in a future contribution. You can however check it works as
expected by dropping the set of seen username per project and check
that you get a cache miss:

   DEL apache_svn_project_set_<project_id>

While this contribution does not add a new feature, the error introduced
by commit 26cbef9a769815c23a2503826e137a1ef5067cb9
("Can't locate object method 'log' via package 'Apache2::Log::Request'")
is fixed.

To test, you need to redeploy the Tuleap.pm file to
/usr/share/perl5/vendor_perl/Apache/Tuleap.pm and restart Apache.

This is part of story #12160: have a clean way to suspend a project

[0] https://github.com/antirez/redis/issues/135#issuecomment-2361996

Change-Id: Ib056f7c3ca1c37bdf39de8c1106ff78e248c9970