A non active project must not be used as a source of ugroup binding
This is part of story #12160 have a clean way to suspend a project

Non active projects were not listed in ugroup binding modal.
Now the backend enforces this verification.

Already existing binding on a suspended project are not removed.
Suspend a project does not remove or update already existing binding.

Change-Id: I01e188c362df771d0c9951b85bbf8c6721223b25