General async logs should not go in the logs dedicated to mail information
The approach chosen in 5ee0b67774dbd95d36116bc64f41427f6b782e1e was not correct
as it redirects general worker logs to the mail_syslog file which is not wanted.

This contribution reverts the work that was done and ensure the tracker mail
notification task uses a logger dedicated to mail information.

Part of request #12435: We should log more information when mail trigger an exception

Change-Id: I96114a7126650970cafa8850bd85b40771aff7ce