No more able to link an artifact located in a non active project
This is part of story #12160 have a clean way to suspend a project

When an artifact id matching an artifact located in a non active project
is provided in the artifact link field, the value is no more considered
as valid and an error is displayed.

In addition, updating a project status force the cache to be regenerated
for this project.

Change-Id: I754bd287e2626799424ba7b99d7b57929b3ee225