Glossify error states
part of story #12160: have a clean way to suspend a project

Go to:
- a suspended project
- a private project you are not member of
- a public project you are not member of, and you are restricted user

Margins should be the same.
Buttons should be large.
The page displays the latest versions of illustration.

The footer should not be glued to the buttons in IE11. The fact that the footer
is not at the bottom on IE11 is another issue, see request #12493.

Go to tlp documentation » visual assets » colors.
You should have explanations about illustration colors usage. If you switch the
theme color with the tlp selector at the top left, then the colors of the svg
example, and the hexa code of associated colors are changed accordingly.

Change-Id: I6e3a6d9f0c46af3e35943b7de61a3fb747be849d