Use latest version of tumbleweed
Part of story #12160: have a clean way to suspend a project

This should be the last illustration upgrade for this story.

Go to:
- a suspended project
- a private project you are not member of
- a public project you are not member of, and you are restricted user

Margins should be the same.

Go to a profile page of a user that isn't member of any projects. The
used illustration should be the latest one.

Change-Id: I178fb5fa4dde9250dc579e8a300254ecf68d912b