request #12515: Remove all usages of the mysql_* API
The usage of PDO as the main way to access the DB can be
considered safe enough to drop the fallback to the mysql_* API.

The DataAccessResult class is kept as an implementation of
LegacyDataAccessResultInterface for test purposes to avoid a
rewrite of quite a few existing tests.

The option fallback_to_deprecated_mysql_api is not available anymore.

Change-Id: I5bb4429a8c908892a44181588d882d308ecfa368