request #12497: Restricted user get Fatal Error on homepage
As restricted user, go to homepage, a user profile, or any page a restricted
user cannot access (depends on the platform configuration). You should get an
error state and not a fatal error.

As restricted user, go to a project you are not member of. You should get an
error state.

As regular user, go to a private project you are not member of. You should get
an error state.

As regular user, go to a suspended project. You should get an error state.

Change-Id: Iaeecd2772565d4a5591ca5c2d2096d38249b3559