request #12513: Commits associated with an empty email author address show a wrong author in the web UI
When Tuleap does the association between an commit author and Tuleap user,
the commits with empty emails should be left out. Empty strings as emails
are in this context not enough to get accurate data.

The issue can be reproduced by creating a commit with an empty email
and pushing it to a Tuleap repo:

    $ git -c author.email='' commit -m 'Commit message'
    $ git push

Change-Id: I623b4fb0ab9bce57efebdfb4d05c8aac990d1c17