Stop loading widget's assets on all dashboards
Part of request #12590 Widgets should not include their CSS files on
empty dashboards

This patch aims to improve performances on all dashboards by removing
unused assets.

How to test:
For each of the following widgets:
- Cross Tracker Search
- Document Manager: Documents under review (User Dashboard)
- Continuous Integration: Test Results
- Labeled Items
- Project Links
- Timetracking Overview
- Personal time tracking
The following is true:
- Their CSS / JS files are NOT loaded on an empty User / Project
dashboard
- When added to a User / Project dashboard, their CSS / JS files are
loaded

Performance notes:

Before this patch, on an empty project dashboard, without any extension,
Lighthouse on Chromium 70 reports a Performance Audit score of 52 with a
Time to Interactive of 6930ms (with simulated 3G and 4x CPU slowdown).
We load 12 CSS files and 5 JS files.

After this patch, same settings: Performance score is 63, Time to
Interactive is 6,370ms.
We load 6 CSS files and 4 JS files.

Will be done later:
- Renderer Widget makes us load Tracker / Cardwall / Graph on trackers
v5 CSS files on every dashboard. It will be dealt with separately.
- Provide a way for widgets to declare javascript dependencies for
adding widgets (for Labeled items to load "configure-widget.js")

Change-Id: Ia66c58147eab7c92b71303ae81164c8041fa992a