Replace spinner by skeleton screen
part of story #12443 Table: Minimal folder display

Instead of displaying a spinner, display a loading state.

Empty state is now displayed in a card in order to avoid too much flickering
(there is still flickering but it is less visible than if there is a transition
card -> no card.

Therefore size of the illustration needs to be decreased and whitespaces
adjusted. This will be done in a dedicated commit.

Change-Id: Ic0218d1a17d7a786dc58a6430d7dffd70bc827b2