Document display minimal table content
part of story #12443 Table: Minimal folder display

This patch introduce the display of folder and items in table.
There is no distinction between folder and docman items.

Folders are not clickable
Docman items are not clickable

Users are properly displayed and are clickable

Last update date is not formated like in mockup, this will be done in a dedicated commit

Change-Id: Id1b26b391b9a2e0bdd86052198a3b33e40bccb1c