Duck typing should distinguish bind types
Part of story #12523 Export single-value list fields to CSV

How to test:
- Export CSV from a cross-tracker widget with a saved and non-empty
report.
- Selectboxes and Radio buttons that have the same name but not the same
bind type (static, users, ugroups) should yield two columns (with the
same name) in the CSV export.
- Count of tracker fields should also take bind type into account. That
means that when a list field with the same name is present in only one
other tracker, if it has not the same bind, it won't be exported.

Change-Id: Ie0413feb5e25cfffaf683e044a49c9eceb851f07