Handle authorization verification when processing Git LFS Batch requests
When no reference is given in a write batch request and fine grained
permission are enabled for the repository, the access is denied.

You can test the implemented HTTP route by crafting manually a
valid Batch request.
For example, with curl:

    $ curl -X POST -H 'Accept: application/vnd.git-lfs+json' \
        -u username:password \
        'https://tuleap.example.com/plugins/git/projectname/reponame.git/info/lfs/objects/batch' \
        --data \
        '{"operation": "download", "transfers": [ "basic" ], "ref": { "name": "refs/heads/master" }, "objects": [{"oid": "12345678", "size": 123}]}'

The specification of the format of a Batch request can be found at [0].

Git LFS HTTP routes are only accessible when an undocumented feature flag
('git_lfs_dev_enable') is enabled until a minimal support of the Git LFS
workflow is implemented.

This is part of story #12322: have git-lfs batch and basic transfer API

[0] https://github.com/git-lfs/git-lfs/blob/v2.6.0/docs/api/batch.md

Change-Id: I7a63fc5ba8df834f89099553a4ffe1a47291d5ac