Update REST route GET /projects to get plugin's criteria
Part of request #12464	Update project rest route to get time tracked project

This is an intermediate step to be able to get projects with timetracking.

The only difference with master is that now the route GET /projects accept a new way to query projects: { with_timetracking: true }
However it only validates the query, the search is not done yet. It will come in a next changeset. Also, it will be documented later.

Actualy, this request return the same result as request "is_member_of". It's will be fixed on the next patch

Change-Id: I75b0880e6a24567de4f7d5228c5874b8b66cae94