request #12549: Password reset procedure uses generic Tuleap split token
The goal is to mutualize code between different parts of Tuleap
using this pattern.

Existing password reset tokens will be considered as invalid,
users will be forced to regenerate a new token after the upgrade.
The token being short lived (1 hour) that should not be too much
of an issue.

Change-Id: Iff427672c825b144745a10a66c228b8e84748738