request #12550: Mark personal access keys as usable in cross origin requests
Similarly to the auth token header, the personal access key header
should be marked as one that can be sent in cross origin requests.
Without that, a web app deciding to manage all the requests on the
client side and wanting to use an access key won't be able to do it.

Change-Id: I8ef1caeef9e92d2cd515d337c02681a86a79b370