Handle authorization of basic transfer upload actions
The object is still not saved but an authorization is attached
to each actions returned by the batch request API. Someone without
the proper authorization trying to upload an object will get a
rejection.

To test you first need to request the upload of an object to the batch
request API and then try to upload it to the given URL with the given
authorization header.

Since the plugin is only accessible in dev environnement and did not
contain any persistant data until now, a migration script for the database
change has not been created, please reinstall the plugin or create the table
manually.

This is part of story #12322: have git-lfs batch and basic transfer API

Change-Id: Ic9ba63202db1305b3c7aea0d223e81e25b7ce216