Git LFS client can be used to upload files
For now, since it is not possible to download them to verify,
after a succesful upload you should find the corresponding information
in the database (plugin_gitlfs_object and plugin_gitlfs_object_repository tables)
and on the filesystem (by default at /var/lib/tuleap/git-lfs/).

league/flysystem is used to abstract interactions with the data storage. For now,
we only store on the local storage but in the future that would allow us to store
the large files to dedicated object store.

This contribution comes with some shortcomings to keep it managable in terms of
reviews. These shortcomings includes:
 * size of the object is verified only after being written a first time on the
disk. This mean a user might upload a larger file than what has been announced
to the batch API. The file is deleted after being uploaded completly if the size
does not respect the announced size though.
 * OID of the saved object is computed by reading the file that has just been
written to the disk. This is a waste of IO, the OID should be computed on the
fly while writting it.
 * nginx configuration might prevent you to upload a "too large" large file
and can be tweaked to be more adapted to kind of traffic.
 * objects that are not marked as available can left files used during the
processing. Notably this can happen if the client never calls the verify
action or if there is a hard failure while the code try to do the clean up.

This is part of story #12322: have git-lfs batch and basic transfer API

Change-Id: I0d00325b019c64095021dd4c233f4cba3a3c125f