Fix Git LFS action content response format
The returned response must respect the format defined in the
specification [0] otherwise the Git LFS client will either reject
the response or ignore the attribute.

In this specific case, the header attribute of actions is incorrect
so the Git LFS client does not add the specified header to its action
queries. Since the header Tuleap adds is needed for the authentication
all action queries done by the Git LFS are rejected.

This is part of story #12322: have git-lfs batch and basic transfer API

[0] https://github.com/git-lfs/git-lfs/blob/v2.6.0/docs/api/batch.md#successful-responses

Change-Id: I1c7486365bd64960e3d6476320f76c6feaed92ec