Extract the LFS Object representation from the batch request representation
The LFS Object representation is going to be needed outside of the
batch request space, this contribution deals with this refactoring.
Also, the LFS object ID format is now enforced to be sure Tuleap does not
try to work with LFS object with impossible OIDs.

Besides enforcing the LFS object ID format, no functionnal change is expected.

This is part of story #12322: have git-lfs batch and basic transfer API

Change-Id: I929c3cacc827dadca22b7fa11a4dc5a4774497a4