request #12562: Default nginx configuration for plugins might serve requests to the wrong plugin
Currently if two plugins starts by the same name the plugin with
the shortest name can serve requests that were destined for the plugin
with the longest name. The issue can be made visible by declaring a route
under /plugins/graphontrackersv5 and having the legacy graph on trackers
plugin enabled.

Also the current configuration is not sane because it potentially exposes
to "nginx off-by-slash" kind of issue [0].

[0] [PDF] https://i.blackhat.com/us-18/Wed-August-8/us-18-Orange-Tsai-Breaking-Parser-Logic-Take-Your-Path-Normalization-Off-And-Pop-0days-Out-2.pdf

Change-Id: Ic554c1f0644498603fcffce9b3627bb55e982f60