Git LFS operations can be authenticated from a SSH connection
With this contribution it is not mandatory to use the HTTP
Basic Authentication to do LFS operations.
The authentication process from an SSH connection is described
in the Git LFS specification [0][1].

To be able to test this, you need to deploy the git-lfs-authenticate
command as described in 0a7cd6ee39f28cfa308fc55d088cb391070ba03e.

This is part of story #12324: use git-lfs with ssh transfer

[0] https://github.com/git-lfs/git-lfs/blob/v2.6.0/docs/api/authentication.md#ssh
[1] https://github.com/git-lfs/git-lfs/blob/v2.6.0/docs/api/server-discovery.md#ssh

Change-Id: I97135ab18b792d1c74f3eecd63fa74be92c67a1e